rename file, clean up file
