--- params.conf
@@ -1,4 +1,0 @@
-[DEFAULT]
-
-scenario = 1 # 1 for normal, 2 for DDos, 3 for security_enhanced
-time_til_delete = 60
