+++ slice.conf
@@ -1,0 +1,3 @@
+[DEFAULT]
+
+scenario = 1 # 1 for normal, 2 for DDos, 3 for security_enhanced
